Collect multislit tests into test_multislit
